chore: extract ticket-validator sizing helpers to sibling module
Splitting the three-layer sizing model into ticket-validator-sizing.js
keeps ticket-validator.js under the maintainability ceiling. Behavior is
unchanged — re-exports preserve the legacy import surface.
